ci: remove fmt nightly
rust-fmt nightly can refactor examples inside documentation, but it's
more than one month that is missing from the build. This means that the
circleci script will fail and this defeats the purpose of the CI.
For this reason, it has been reverted back to the stable for the CI
check.