feat: add error tracking float
This was not originally present in the C++ version. And turned out to be
a HUGE technical debt: in the end the comparison were being done with a
tolerance of 1E-2 which is kinda crazy, anything lower would result in
strange self-intersection errors.

This technique is straight up taken from pbrt-v3