refactor: vertex buffer in shape signature
I had to change the signature of the traits Shape and Accelerator. The
reason is that due to how lifetimes work in Rust, I can't store a
VertexBuffer and a Pointer to the buffer inside the triangles and move
the buffer as I please. So the pointer to the buffer must be passed on
demand to the various intersection routines.

This complicates a bit the signature but overall is more efficient, as I
know for sure that I need to unwrap it only for triangles and it will
always be available. Moreover i save 8 bytes for each triangle (I don't
have to store the pointer to the buffer in the triangle itself).