refactor: two different shape trait
Backed up a bit on my previous refactor. The reason is the fact that
triangles are not public and exposed only through Mesh. Having a trait
that requires an additional parameter, internal to Mesh, only for them
is not a good idea. Also because an user is never supposed to pass this
extra parameter and always fill with None.

For this reason I created two traits called BufferedShape and
BufferedAccelerator, that work ad hoc for triangles. They have been
implemented using macros to avoid duplicate code.