refactor: finer grained device queue usage
Before this commit a single graphics queue was used. Now a device
allocates a graphic, a transfer and a compute queue and uses the
transfer for loading operations and the graphic for draw operations.
Queues are selected such as that a family never runs out of queues
(looking at you intel that you give me a lot of families each one with 1
queue with all capabilities).
Technically I could use more than one of each kind, but practically due
to synchronization I don't need to. So the device allocates a single
queue for each type.