feat(ui): save scene
Now changes to the camera and the materials (the only possible changes
right now) can be saved to file!

Fixed also a bug that slipped through testing in the previous commit:
saving a scene would destroy the header, hence reopening was impossible
(in the tests I wrote I always used the already open scene and the
contents were correct). Added a regression test for that.