refactor: actual RAII
Huge refactor: moving from a memory model similar to C (due to the
vulkan library being in C), to a proper RAII model with automatic
deallocations when the object goes out of scope. This results in a small
overhead on memory (something like 8 bytes for each allocation) and on
setup/teardown time (mainly because the resources are tracked using Arc
and Mutexes).