build: shaderc can now include files
Despite supporting "#include" directives, google's compiler needs a
callback to actually handle the directive.

REALLY GOOGLE? Who the fuck is the 200 IQ big brain that decided
"fetching the include content should be responsibility of the user"?

What is the next level, a compiler that gives you the compiled file name
and expects the assembled binary from the user?