feat: "meta" chunk in the parser
It does not store metadata per se, but unstructured arrays of ints and
floats, so it can contain virtually anything. Up to the user deciding
what represents each int/float.

I need this to store some additional variables for the scene (like
cleanup color, scene radius, skybox texture) that does not fit well in
any of the other chunks.