feat: glass material
Without caustics, as those are not supported with the current rgen
shader.
The light path is completely skipped for specular materials, and even if
this wasn't the case, the occlusion tester does not account for
transparent materials. Probably I will add caustics in the bidirectional
methods (BDPT and MLT, if ever I will be able to add them, my 3080 is
already suffering with a 5 step path tracing).