Revert "feat: frequency based push constants"
This reverts commit f84c24b1d6b0883978491cc6c56ba5885dd6a0eb.

No idea why, but if I move the order of the (set0, binding0) variables,
the traceRay for occlusion rays completely fails. It is like the call is
never executed. I spent more than a week trying to figure out the
problem (the last few commits are from some days ago but I hoarded them
for a while), and at this point I give up. Tried different alignments,
  different location, different buffers, to no avail.

  Given that I will probably ditch the path tracing in favor of the
  bdpt, I wasted too much time on this silly refactor.