refactor: empty scene is default scene
Instead of checking everywhere if the scene exists prior to doing any
operation, now the scene ALWAYS exists but it's empty.

Minor inconvenience, I have to create a lot of buffers of size 1 byte on
the GPU, because 0-sized buffers are not legal (and neither are 1 bytes,
probably depends on the GPU alignment).