perf: async path tracing
The various pixels are now asynchronous. A pixel may be at step 0, and
the adjacent one at step 3. This means that every pixel is independent
and there is not a "global" state for the current image. This also means
that every trace ray call, draws something to the image (unless there is
no light in the scene).