refactor: queues inside command managers
A command manager was associated with a single queue in any case. Never
saw an error because in the testing system GRAPHIC, COMPUTE and TRANSFER
are all in the same family, but there was an instance in which I
declared a GRAPHIC command buffer and submitted to a TRANSFER queue.

In addition, now submitting a queue requires passing through the Device
interface. This will be useful in the next commit in order to submit
only from the main thread.

Due to the new privacy rules, Device and Swapchain are tightly coupled:
Queue has some private fields and to avoid calling VkQueueSubmit from
outside the Device interface these fields need to stay private.
However, Swapchain also requires to use one of these fields, so it has
been moved inside the Device module (in its own submodule).