fix(ui): camera moving when loading/renaming
Input handling was not exclusive, so the camera was moving when pressing
WASD while renaming an object.
This commit fixes this problem.
In addition, the camera has been locked while a new scene is being
loaded. Not a technical limitation, but the screen is 'overlayed' so it
looks like the movements should be blocked.